[backend] Fix delete indexed file (#11260)
