+++ opencti-platform/opencti-graphql/src/manager/activityManager.ts
@@ -19,14 +19,13 @@ import conf, { booleanConf, ENABLED_DEMO_MODE, logApp } from '../config/conf';
 import { INDEX_HISTORY, isEmptyField } from '../database/utils';
 import { TYPE_LOCK_ERROR } from '../config/errors';
 import { executionContext, REDACTED_USER, SYSTEM_USER } from '../utils/access';
-import type { SseEvent } from '../types/event';
+import type { ActivityStreamEvent, SseEvent } from '../types/event';
 import { utcDate } from '../utils/format';
 import { listEntities } from '../database/middleware-loader';
 import { ENTITY_TYPE_ACTIVITY, ENTITY_TYPE_HISTORY, ENTITY_TYPE_SETTINGS, ENTITY_TYPE_USER } from '../schema/internalObject';
 import type { AuthContext, AuthUser } from '../types/user';
 import { FilterMode, OrderingMode } from '../generated/graphql';
 import type { HistoryData } from './historyManager';
-import type { ActivityStreamEvent } from './activityListener';
 import { BASE_TYPE_ENTITY } from '../schema/general';
 import { elIndexElements } from '../database/engine';
 import { getEntitiesMapFromCache, getEntityFromCache } from '../database/cache';
