+++ opencti-platform/opencti-graphql/src/manager/clusterManager.ts
@@ -17,17 +17,9 @@ import { ENTITY_TYPE_SETTINGS } from '../schema/internalObject';
 import type { BasicStoreSettings } from '../types/settings';
 import playbookManager from './playbookManager';
 import { getAllManagersStatuses } from './managerModule';
+import type { ClusterConfig } from '../types/clusterConfig';
 
 const SCHEDULE_TIME = 30000;
-
-export type ClusterConfig = {
-  platform_id: string;
-  managers: {
-    id: string,
-    enable: boolean,
-    running: boolean
-  }[]
-};
 
 const initClusterManager = () => {
   let scheduler: SetIntervalAsyncTimer<[]>;
