+++ opencti-platform/opencti-graphql/src/modules/indicator/indicator-domain.ts
@@ -50,7 +50,7 @@ import {
   type DecayLiveDetails,
   findDecayRuleForIndicator
 } from '../decayRule/decayRule-domain';
-import { isModuleActivated } from '../../domain/settings';
+import { isModuleActivated } from '../../database/cluster-module';
 import { stixDomainObjectEditField } from '../../domain/stixDomainObject';
 import { checkScore, prepareDate, utcDate } from '../../utils/format';
 import { checkObservableValue, isCacheEmpty } from '../../database/exclusionListCache';
