+++ opencti-platform/opencti-graphql/src/types/clusterConfig.d.ts
@@ -1,0 +1,8 @@
+export type ClusterConfig = {
+  platform_id: string
+  managers: {
+    id: string,
+    enable: boolean,
+    running: boolean
+  }[]
+};
