+++ opencti-platform/opencti-graphql/src/types/exclusionList.d.ts
@@ -1,0 +1,6 @@
+export interface ExclusionListCacheItem {
+  id: string
+  types: string[]
+  values: string[]
+  ranges?: number[] // only used for IPs
+}
